zipFile repository client
