+++ api/core/use_case/export_use_case.py
@@ -18,7 +18,6 @@ from core.service.document_service import DocumentService
 from core.shared import request_object as req
 from core.shared import response_object as res
 from core.shared import use_case as uc
-from core.use_case.create_application_use_case import zip_package
 
 
 class ExportRequestObject(req.ValidRequestObject):
